Move to ArisBasics.h
