+++ common/code/CommandBuilder/CommandBuilder.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "ArisConstants.h"
+#include "ArisBasics.h"
 #include "commands.h"
 
 namespace Aris {
