+++ sample-code/vc-using-framestream/vc-using-framestream/AcousticSettings.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ArisConstants.h"
+#include "ArisBasics.h"
 #include <cstdint>
 
 constexpr uint32_t kFirstSettingsCookie = 1;
