+++ sample-code/vc-using-framestream/vc-using-framestream/ArisBeacons.cpp
@@ -1,5 +1,5 @@
 #include "ArisBeacons.h"
-#include "ArisConstants.h"
+#include "ArisBasics.h"
 #include "UdpListener.h"
 #include <iostream>
