+++ sample-code/vc-using-framestream/vc-using-framestream/vc-using-framestream.vcxproj.filters
@@ -78,6 +78,9 @@
     <ClInclude Include="..\..\..\common\code\CommandBuilder\CommandBuilder.h">
       <Filter>common/code/CommandBuilder</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\common\code\ArisBasics.h">
+      <Filter>common/code</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="generated\availability.pb.cc">
